feat: selecting a task runs it/jumps to buffer
Need to get a hold of compilation mode a bit better so that the fabb
window is reused and/or fabb status should refresh when a new task runs.

This uses text properties to pass data to the select-line command, which
is a bit unwieldy. Probably could instead use a hash-map from line ->
data on the local context var, similar to the rofi api i use in clawe.

Ideally we'd know if the compilation/task buffer is still running or
not, and use/display that in the status buffer's task-line.