feat: reinvoke a task with projectile-completing-read
Introduces a simple way to reinvoke a task, passing an arbitrary project
file as the final argument. This covers alot of ground for me in my bb
tasks, which often expect the same (eg. `bb test-match <file-path>` to
run a single file's tests.)

Uses projectile for now, but it'd be a simple dep to remove, i think
there are generic versions of these funcs.